Move input code to input/ and add public headers to include/littlehell/input
