+++ src/i_system.c
@@ -35,7 +35,7 @@
 #include "m_argv.h"
 #include "m_config.h"
 #include "m_misc.h"
-#include "i_joystick.h"
+#include "littlehell/input/joystick.h"
 #include "i_sound.h"
 #include "i_timer.h"
 #include "i_video.h"
